% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{AIC.wnpmle}
\alias{AIC.wnpmle}
\title{AIC for wnpmle objects}
\usage{
\method{AIC}{wnpmle}(object, ..., k = 2)
}
\arguments{
\item{object}{A \code{wnpmle} object.}

\item{...}{Ignored.}

\item{k}{Penalty per parameter (default 2 for AIC).}
}
\description{
AIC for wnpmle objects
}
\value{
A numeric scalar giving the AIC value.
}
