% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.wnpmle}
\alias{print.wnpmle}
\title{Print method for wnpmle objects}
\usage{
\method{print}{wnpmle}(x, ...)
}
\arguments{
\item{x}{A \code{wnpmle} object.}

\item{...}{Ignored.}
}
\description{
Print method for wnpmle objects
}
\value{
Invisibly returns \code{x}.
}
