% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnpmle-package.R
\docType{package}
\name{wnpmle-package}
\alias{wnpmle}
\alias{wnpmle-package}
\title{wnpmle: Weighted NPMLE for Recurrent Events with a Competing Terminal Event}
\description{
Implements the weighted nonparametric maximum likelihood estimator (wNPMLE)
for the marginal mean function of recurrent events in the presence of a
competing terminal event. Two semiparametric transformation models are
provided: the Box-Cox transformation model and the logarithmic
transformation model. Both use automatic differentiation via TMB for
efficient and exact gradient computation.
}
\details{
The main user-facing function is \code{\link{wnpmle_fit}}.
}
\references{
Bellach, A. and Kosorok, M.R. (2026). Weighted NPMLE for the marginal mean
of recurrent events with a competing terminal event.
\emph{arXiv preprint} arXiv:2605.25934. \doi{10.48550/arXiv.2605.25934}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abellach/wnpmle}
  \item Report bugs at \url{https://github.com/abellach/wnpmle/issues}
}

}
\author{
\strong{Maintainer}: Anna Bellach \email{abellach.biostat@gmail.com}

}
\keyword{internal}
