% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{BIC.wnpmle}
\alias{BIC.wnpmle}
\title{BIC for wnpmle objects}
\usage{
\method{BIC}{wnpmle}(object, ...)
}
\arguments{
\item{object}{A \code{wnpmle} object.}

\item{...}{Ignored.}
}
\description{
BIC for wnpmle objects
}
