% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.wnpmle}
\alias{coef.wnpmle}
\title{Extract coefficients from a wnpmle object}
\usage{
\method{coef}{wnpmle}(object, ...)
}
\arguments{
\item{object}{A \code{wnpmle} object.}

\item{...}{Ignored.}
}
\description{
Extract coefficients from a wnpmle object
}
