% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.wnpmle}
\alias{logLik.wnpmle}
\title{Log-likelihood for wnpmle objects}
\usage{
\method{logLik}{wnpmle}(object, ...)
}
\arguments{
\item{object}{A \code{wnpmle} object.}

\item{...}{Ignored.}
}
\description{
Log-likelihood for wnpmle objects
}
